services/detection/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
joblib==1.3.2
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7
httpx==0.26.0
requests==2.31.0
